Get more releases than tags to handle wrong message
If we get equal amount of tags and releases, there was a
case where a tag would be shown as not released.

This happens when when a release is created later on a tag.
So if we fetch 3 tags and releases where total tags are 5.
1 2 3 4 5 are the tags in order of time created (1 is created
latest)
And 1 5 4 2 3 are releaes i.e. 1 is released latest.

When we fetch we get 3 tags 1 2 3 and 3 releases 1 5 4 which
would lead us to believe that 2 and 3 were never released.